Update README.md
I revised the explanations for bubble sort and insertion sort. There was a mistake for the bubble sort in which the bubble sort starts from the rightmost value which means the value that it is comparing to is actually to the left. Hence, going down the elements is actually moving to the left (not right).

Refined our explanation of partition in Insertion Sort. Came up with a nice nickname for Insertion Sort and called it an "Invisible Wall of Partition", thereby not shifting the indices to the right by 1 while inserting the partition.